requirements.txt
psutil
requests
flask
mysql-connector-python